Split the constants to another file.
